feat: moved org from templates to features
